3.1.12 changed Node to not public and merged with StackList class
